requirements.txt
boto3==1.9.139
botocore==1.12.139
docutils==0.14
future==0.17.1
jmespath==0.9.4
numpy==1.16.3
python-dateutil==2.8.0
python-dotenv==0.10.1
s3transfer==0.2.0
six==1.12.0
tableprint==0.8.0
urllib3==1.24.2
wcwidth==0.1.7
